Read configuration from default_config.yaml and config.yaml
We are going have secrets (passwords, OAuth keys, etc) and we do not
want to put those into version control. Add support for reading YAML
configuration files to load configuration settings that we can use
elsewhere in the application.

Note that we are also adding the optional config.yaml file to our repo's
.gitignore file so that we do not accidentally check it into version
control.