Use yaml.safe_load
The `load` method of the Python YAML bindings allows the
`!!python/object` tag which could be used to inject unexpected state if
an untrusted user modified the configuration. This is not a major threat
in typical configuration file parsing, but this tutorial is in part
designed to promote best practices, so we should showcase using the
`safe_load` function which is restricted to standard YAML tags and
explicitly whitelisted objects.

h/t @legoktm